feat: separated player data to team component
