+++ src/app/components/Team/Team.jsx
@@ -1,87 +1,15 @@
 "use client";
 
 import React from "react";
+import playersData from "../../utils/playersData"
 
 export default function Teams() {
-  const jugadores = [
-    {
-      nombre: "Bruno Garay",
-      posicion: "LW, MI",
-      equipo: "Equipazo FC",
-      edad: "21",
-      nacionalidad: "Uruguayo",
-      dorsal: 18,
-      imagen:
-        "https://lh3.googleusercontent.com/CIHrlVodv8Fe51Q-MTbGoUj1-CGPZ7Lm1SeY7zVq9hxBKg_J2zMDkxnXUs3a-A54FP3-yjmHS5LF7GwyJ3OXOCng1iAXUdbNKFCAOkE2Kn_iB-CDMfl6pQJY1o54HL5GDvzk5rAFMsA7fH2i7M3opRgrXiktXI5LdJ25hVhMBwyR3s7u00FT6FEXvZUGgO-b5UoQ74oBjGbeD6Cux3bZQ5OA3Re3v_uqB1vtjkNuz88RYtAJTtT89p1kM0Ae3EF9bUYeT1NtrR5J0NW_AtLfN6MABqTED3cdsSyr4DltG1hKqxPRzVsfyh3mOkDvPbvWifqnXAHWuTLWtiVZggc6JYcr30Kyc39vazDJNLvkJfMQxpe9qmfp-VmcjYairJ0CbaReKhsCkdz6e9bIp_mpkQ0Okl3Bd2Hz3u__F_INZLtOWidAkb1_6HBmdu1KGoCJWnSnIauLKf5a6qZ8_9jAWKod1EZKBBfWGt5TicTV8lhF1MBCB6WCR2sPws4yu-rCAmkjUPzp3Cn7BKRMP1WgWKiZw17amE2fQ96GYDoXvStuAUF4MB2iAAR79598WFxwEO_RbUW-Ca0TeLnHoueUDH6enSkSIrShlOu-JFIKrwqa45A07yTbVcOU7NYtxD3IyT_CfUp6z3_yiflc2Ep5R4l0Y8CeGJV67NMTB9zxd1zBmhBpZD9Kcj3aL6ypBQP_5Uu6-78avi3050NSYws1xsYEXQwVJXWVDwTyZnmPuAfXtP12GFim-P_FGGThF42muh5fiRvAV2IWev7eV5lSMLyqPdbmcDh14wItrjEC3l9tjhRUyTwQ3_Vi0BstXUl-3kyqpL_I5hwFFl0aFogoN20zPgW5t3mW0c7BImpuRBMzgHuZbYiNWBlxnhJIGtontlVr3T92_wX_S-_kWnHy8zHwuxXqgl-vP9fvFg2D0cDD0iduF8g5AvDHG2w60hiMvRU68JtdaTD3_RvUkrqy=w525-h293-no?authuser=0",
-    },
-    {
-      nombre: "Pablinho Méndez",
-      posicion: "DC, MCO",
-      equipo: "Equipazo FC",
-      edad: "20",
-      nacionalidad: "Uruguayo",
-      dorsal: 10,
-      imagen:
-        "https://lh3.googleusercontent.com/ntpgcnejonsR4fqh7Ll2vFcfQ2r8w918vIvpa_ffy47427oht3VxI-P28MzdEVKT1s9mxQ1JM7ld_PRKrj_Xu7dlLLMwzMn0j0dLXoA2Njpx_LQbY2T9hm1o8a_9CTMjE0EIi4MjHO1simLdj5eTUEtwJC7p5sg9t4ZsXmu8-fKjNgFXmQx-MWFdTbbCzTC1gURwo6LHlDhtAouS6of7y_VBx6Vj1Q0lmZwfKbhdceMmBABrhybXHQt3vRlLQw66MaQECLaUZZAnB0rAkwwEy2PH9Yy-IsdD9w_9q4q1W76_PeFLs6HecGMYHLaoLHgGKMpJ-m2DsTve4_uGVkK_aZTcVhlOEtFke7YU1qa4oMYIcwYdOtHRWi2yrWWtOvxBgVSPv1bM4P7I-zOh6QeyZgLTGl9HSyJuA0q5hJHvlWFhj7OH5YJYEK-Rmm2pv_6PDwdNLep0N2ANZJql-7ePhtM5XbjT3ARg79yA-587UKveLQ30T1nTG70Fkk-9I5zNplCcS1HoxJjCVG1N0xoUS2w1F4nHk3JKu_9H8q4L2Kvj9A8aVgmOZHiTjM5AFuPps4MlvfK-8QX7HkHFO_ARjv0EPhSk7Jj8CehBvhJI-01NYYSAcBt7USHnKQg4fTSIgZrPNCg6-wL-L9l2HDgBAN1aCmxElenNZsLt--pnPTAxy4V-nMl70lWNEnlLlSC21hbDfXc0RHHAlc4R2VSrGxgcFKYg3lhBwHAdMxU7IzOrVADS0F8A5nl4QcWjQF_Qbb69IMS4bPt9Op3LZp3HdVBOHGwgvcNK3Dp5GZE_yBJrrHlnlt95qlqB3IIFD0EFmQa6iUF9-zax26jQq2eoehB91Hhw6MoY7FRHtYHjbKZNFj-mpHEC3Wou55SKa4wufhgj2_eDP2uPZqfhfn0YIOB5c2LiKl-05sxK9BUDGxgTvjB5k5y4M_ZPC7bG8B1wVfwEscJWc93pjz-0ers5=w1168-h657-s-no?authuser=0",
-    },
-    {
-      nombre: "German Centurión",
-      posicion: "MC, MP",
-      equipo: "Equipazo FC",
-      edad: "20",
-      nacionalidad: "Uruguayo",
-      dorsal: 22,
-      imagen:
-        "https://lh3.googleusercontent.com/wPQ8inVouy04InJc9a9O2MR_-9rPPcHd1ioK7IbmMVomwCz_yAM0b4npnXv-38MYCEaA4-A4lSO5k9DBd9wRgHJ9murFSD8FwKAruifTvsjp3GQlV6PLcv2yDbpKhKiHQNs9YSSWzMrBhsDwoBfxHu64yBkBMm-1LJxeQ_rRpnO4rjkEDTxsgPlg7015aqUjF3AVN9Lt6KxAGRAzkPZ4-avtWPwTTifRxBrMaqG5iPJCttnLcp7uPIv99GEaUeIjVfmAxdDHK_HqcYG7-DW9a-DN0e8VATswD7VlLzI21RXKx47MuSHzNH-kGWhbWGxNBpUHMR-mCiHEcavI8jb4Zb0HNR3cm_UbtKKOfwf1plmI5m_AO2fsHcPFj22TLFKe1m5FFMLu_BidRqEC6mGkqvjcWqB09ZF10VASt4ROo50QANLHE-MJl6Yj-FBl6vRyUy0_Zoq5XFOLdzxw7lLHtv5mSs9GEIQBHLTLL6hRjS2y6PcBhsxrbV1PcPk6EJd6FdO4PwP_JzYPOO8sFUpNZ7PxPVv2mrkALwoklQeifI7WnqOlXVx7prdfUOtOy0-Pb-PhNwzFZwdfb17pH1eLimDQq4HHFL8Pz8SD4QPGGXNLB4R69jOVb8OjdKUIx_B2z7og5bIFpdxLVhMUcKrlHFuBJ2B_kdYqFSWd4okL8hSNWKdOAIO7LEmmm1qrrYPt8c8jHwRYpX97vKionlEV2BPHJOVYmtk_OSlB9EIX8XNhZBLy2ntFV09E6c2-br8U709dpJskaOsfB_rw9EasV2CDw1G7htGV6YxjAMBjfbQ-jHT9lMBQEAndzsmBIaKKeX_-bn5ZN-kbBmIuZlwaEl1i4KMT6pFTNiereh0i6mb5IaeLsNX2ZWoiD9LAyVO9yZURjJTMyb00CB7gOLTdN-qoc2eISe18tGdobxk5c8lhLJj7yKxu3ZkAbio20fJtIiSoNyB2uSxf-SEOocga=w1168-h657-s-no?authuser=0",
-    },
-    {
-      nombre: "Maciel Galeano",
-      posicion: "DC, SDD",
-      equipo: "Equipazo FC",
-      edad: "21",
-      nacionalidad: "Uruguayo",
-      dorsal: 10,
-      imagen:
-        "https://lh3.googleusercontent.com/bwUpiLactSNAK_59Ro2To-RY6M91jv6RcZYwJLKIt3lHyLmyJdZSkdGc87lKP5L-uuwS5t10SP0ZAmjFJ9OsP7cwsMWGiiv0zjSPUg_Ykhz_yYCLgnYyDGnulhnqW5iJ3GX8_Q7xQiNa_8JGdmWyTOEyqvFp54VlUIXdNwwjogKjJiHr4YKL-KHsgW0BoY5-h5pTUaihF_xWlSWGRClZFPfACEfbcuVi1PmmuVoUnO2BicHiY6jqtmKdkG5HlVkYlGde2CH9pSV-pdjFvBJUIHzCTqV7fD1jKoYK8hP474mrKPq8f4ngKnfnGmavtxjDklhCYQ6ClxWy2-IF7msyXiRTh5zuJpF_hsTYcAy-hUEfIdT5hDs90QR_xVBaXqFZR9zxGb4J9YruopxfVnleFuAXU8g-eFnEzVhS1s90irpc9__mJbgQB194NLJouZtuoEw8eK1VysGP30Oma-emfpNs_EYBOqGaLFN9-6WC3PdC8efBI-Qp5FFHu3weW2LTz-XGGwHZGc0lC05O8xKpg3mCoSpy9tQTfVlcjPSU11G9ENDjmlHnZGEDE69EBcsPQFEGuAMtcCm8a8eXr5zJ0hZLZuzzt0hhlTvxFNUIM_zbsTcmQ_zFj7alqEJe9Wn02s5nXnfsVOAVPBSTqFMZ2GzesQYwBYE6HsIfMFtNkVz7npjSVNlu8PDtnz_SSWQ2f41876fZGzMzEnZTw5xvRNr9AJzgYnz1tDPkqKVXoTmCe58SBcn1bHWTPKsbML5nxbeZqAPWAgEL6-EVkqUJ1H99eecby8kc3DMgKaSe-6pPl6Qg8gvz5PxuQI3zix8jrNXIxe33ccFZeGMvizVSvZ-fhchZ3y8yMuMU0wfil3OULPz_ViE-FMPIu_EqPghtTYukeGgoVbZrWuB2REMY7tP1Rmx4uwnOCxb5zHYWaTVKwzNIG9ItQ4DkQ3w6wDPUZ1sMo24Izpim0oj7BZ7n=w584-h657-s-no?authuser=0",
-    },
-    {
-      nombre: "Ignacio Lima",
-      posicion: "MC, MCD",
-      equipo: "Equipazo FC",
-      edad: "21",
-      nacionalidad: "Uruguayo",
-      dorsal: 5,
-      imagen:
-        "https://cdn.vox-cdn.com/thumbor/-96l2N7Ucf27ngOVwbn-YZA2fCU=/0x0:3211x4282/1200x800/filters:focal(1472x594:1984x1106)/cdn.vox-cdn.com/uploads/chorus_image/image/71895668/1246457044.0.jpg",
-    },
-    {
-      nombre: "Franco Méndez",
-      posicion: "RW, MC, MCD",
-      equipo: "Equipazo FC",
-      edad: "22",
-      nacionalidad: "Uruguayo",
-      dorsal: 7,
-      imagen:
-        "https://lh3.googleusercontent.com/xJSE6j2PolMHsyQN_MgVUv58gG_EiSW3420DrL4Gy5vwSXXtcWqOwmz7W17Ml6WJIYzkTpNqcAG-ebdhLW75ECyUtFR4NQSoqXP_QX2lnbUYCdJTwRg1EWoHp4zz1NxuAtoCwTe3MSXab9YbDThgysUopLWbPcvZsPYx6p1Hc7jAEQ77nvoWPlTsbmDkGg0Y9CABnE1Eh3lNPlRengbLrhOJbLxj0zT2nhqXTZZH29nlbjAD0Y8PqEHVlPCAdbwOBMpHg3_0FBoxuJdpcyl3ZnwW737FyWdTBF7FYZuJaTIx05rdlERjPyKFJ9ddLm9ZaaIHDlpbOorCUyVMUNNKD_PCgCHWrKxQdfy2PuiYmQgw_LB8cn3rLcLepV6K9YFX69fWa5qP7pCU2-boT3923o7lHOXgXUPcRFyLovcn5b5K75S9XoSiXHSzEbVpZL1dk2hXkKIST03CoeIBumIj1i0gLjDTTdbRkWOe63cxV43LtqG_XUgJihFd6eg6CxqgQnJDh9DEpx2ydrhlvQA0NFBZdEN48NBU0_ZWZi1q2ToPsjaXvyMCZ2MOx8bsd8DqZ6FigKioBzKoZCdGwj4VHQO_d57koV90vo0hX3amkHFX9h1i98HndX6L9u-wenJfYyHaQV912DignqW4X7Wl70izifJHRxeMjIu26hoYxIY-e8VLWuU2hy9T4m9ucX6lNFqRkB9qLitoEisfB2vGQ8S68k3DCohqpGEFXKaFR2GNVUy5Cq3lGqXBsw6-3iwh09asBww-Q9Kj4WPkfOnJajx8XZArUuQp_VozvkE0X9ccjFDPxL-DG__dSrghCDmUfxqRPTEOrLFgn7B87DhcBKxGqdKODkPCk3XTgw2GRF_Jl6lRnBhrfH0ojgrEusA6P1RkFRA0Xqj-E_A_Z5uMjyVVveShNvTK6Oa_zKARVdSv3G2SCuBEK03IMYWWFPH5dQLYq6CB790EtusKxuV2=w1172-h657-s-no?authuser=0",
-    },
-    {
-      nombre: "Manuel Perera",
-      posicion: "MCD, MCO, DC",
-      equipo: "Equipazo FC",
-      edad: "21",
-      nacionalidad: "Uruguayo",
-      dorsal: 20,
-      imagen:
-        "https://lh3.googleusercontent.com/CIHrlVodv8Fe51Q-MTbGoUj1-CGPZ7Lm1SeY7zVq9hxBKg_J2zMDkxnXUs3a-A54FP3-yjmHS5LF7GwyJ3OXOCng1iAXUdbNKFCAOkE2Kn_iB-CDMfl6pQJY1o54HL5GDvzk5rAFMsA7fH2i7M3opRgrXiktXI5LdJ25hVhMBwyR3s7u00FT6FEXvZUGgO-b5UoQ74oBjGbeD6Cux3bZQ5OA3Re3v_uqB1vtjkNuz88RYtAJTtT89p1kM0Ae3EF9bUYeT1NtrR5J0NW_AtLfN6MABqTED3cdsSyr4DltG1hKqxPRzVsfyh3mOkDvPbvWifqnXAHWuTLWtiVZggc6JYcr30Kyc39vazDJNLvkJfMQxpe9qmfp-VmcjYairJ0CbaReKhsCkdz6e9bIp_mpkQ0Okl3Bd2Hz3u__F_INZLtOWidAkb1_6HBmdu1KGoCJWnSnIauLKf5a6qZ8_9jAWKod1EZKBBfWGt5TicTV8lhF1MBCB6WCR2sPws4yu-rCAmkjUPzp3Cn7BKRMP1WgWKiZw17amE2fQ96GYDoXvStuAUF4MB2iAAR79598WFxwEO_RbUW-Ca0TeLnHoueUDH6enSkSIrShlOu-JFIKrwqa45A07yTbVcOU7NYtxD3IyT_CfUp6z3_yiflc2Ep5R4l0Y8CeGJV67NMTB9zxd1zBmhBpZD9Kcj3aL6ypBQP_5Uu6-78avi3050NSYws1xsYEXQwVJXWVDwTyZnmPuAfXtP12GFim-P_FGGThF42muh5fiRvAV2IWev7eV5lSMLyqPdbmcDh14wItrjEC3l9tjhRUyTwQ3_Vi0BstXUl-3kyqpL_I5hwFFl0aFogoN20zPgW5t3mW0c7BImpuRBMzgHuZbYiNWBlxnhJIGtontlVr3T92_wX_S-_kWnHy8zHwuxXqgl-vP9fvFg2D0cDD0iduF8g5AvDHG2w60hiMvRU68JtdaTD3_RvUkrqy=w525-h293-no?authuser=0",
-    },
-  ];
-
   return (
     <div className="container mx-auto" id="teams">
       <h1 className="text-3xl text-center text-red-500 mt-20 mb-10">EQUIPO</h1>
 
       <div className="flex flex-wrap justify-center gap-8 p-8">
-        {jugadores.map((jugador, index) => (
+        {playersData.map((jugador, index) => (
           <div
             className="w-full max-w-lg mx-auto bg-white rounded-lg shadow-lg overflow-hidden"
             key={index}
